# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentExecutionOut": {
        "properties": {
          "agentId": {
            "title": "Agentid",
            "type": "string"
          },
          "calls": {
            "items": {
              "$ref": "#/components/schemas/LLMCallOut"
            },
            "title": "Calls",
            "type": "array"
          },
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageOut"
          }
        },
        "required": [
          "agentId",
          "status",
          "usage"
        ],
        "title": "AgentExecutionOut",
        "type": "object"
      },
...
      "ApiError": {
        "description": "Forma única de todos los errores. La UI la traduce a un mensaje por `code`.",
        "properties": {
          "code": {
            "description": "Identificador estable del error, p. ej. NotFoundError",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Datos extra del error",
            "title": "Detail"
          },
          "message": {
            "description": "Mensaje en español, ya legible",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ChangeMapEntryOut": {
        "properties": {
          "added": {
            "default": 0,
            "title": "Added",
            "type": "integer"
          },
          "change": {
            "title": "Change",
            "type": "string"
          },
          "criteria": {
            "items": {
              "type": "integer"
            },
            "title": "Criteria",
            "type": "array"
          },
          "file": {
            "title": "File",
            "type": "string"
          },
          "removed": {
            "default": 0,
            "title": "Removed",
            "type": "integer"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "title": "Symbols",
            "type": "array"
          }
        },
        "required": [
          "file",
          "change"
        ],
        "title": "ChangeMapEntryOut",
        "type": "object"
      },
...
      "CodeReportOut": {
        "properties": {
          "changeMap": {
            "items": {
              "$ref": "#/components/schemas/ChangeMapEntryOut"
            },
            "title": "Changemap",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "stack": {
            "default": "",
            "title": "Stack",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "summary"
        ],
        "title": "CodeReportOut",
        "type": "object"
      },
...
      "DeliverablesOut": {
        "description": "Un informe por agente. `null` = ese agente no se ejecutó (y `missing` dice por qué).",
        "properties": {
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CodeReportOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "kiuwan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KiuwanReportOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "missing": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Missing",
            "type": "object"
          },
          "privacy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrivacyReportOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "sql": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SqlReportOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "tests": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestsReportOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "uiux": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UiuxReportOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictReportOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "vtr": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VtrReportOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "DeliverablesOut",
        "type": "object"
      },
      "FindingOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "file": {
            "default": "",
            "title": "File",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "safeguard": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safeguard"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "suggestion": {
            "default": "",
            "title": "Suggestion",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "severity",
          "title",
          "detail",
          "source"
        ],
        "title": "FindingOut",
        "type": "object"
      },
...
      "KiuwanDefectOut": {
        "properties": {
          "category": {
            "default": "",
            "title": "Category",
            "type": "string"
          },
          "falsePositive": {
            "default": false,
            "title": "Falsepositive",
            "type": "boolean"
          },
          "file": {
            "title": "File",
            "type": "string"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "rule": {
            "default": "",
            "title": "Rule",
            "type": "string"
          },
          "ruleId": {
            "title": "Ruleid",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "ruleId",
          "severity",
          "file"
        ],
        "title": "KiuwanDefectOut",
        "type": "object"
      },
      "KiuwanReportOut": {
        "properties": {
          "defects": {
            "items": {
              "$ref": "#/components/schemas/KiuwanDefectOut"
            },
            "title": "Defects",
            "type": "array"
          },
          "fileName": {
            "default": "",
            "title": "Filename",
            "type": "string"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "rows": {
            "default": 0,
            "title": "Rows",
            "type": "integer"
          }
        },
        "title": "KiuwanReportOut",
        "type": "object"
      },
      "LLMCallOut": {
        "description": "Metadatos de una llamada al modelo. Sin contenido: podría llevar PHI.",
        "properties": {
          "agentId": {
            "title": "Agentid",
            "type": "string"
          },
          "at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "attempts": {
            "default": 1,
            "title": "Attempts",
            "type": "integer"
          },
          "durationMs": {
            "title": "Durationms",
            "type": "integer"
          },
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "promptVersion": {
            "title": "Promptversion",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "toolsRejected": {
            "items": {
              "type": "string"
            },
            "title": "Toolsrejected",
            "type": "array"
          },
          "toolsUsed": {
            "items": {
              "type": "string"
            },
            "title": "Toolsused",
            "type": "array"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageOut"
          }
        },
        "required": [
          "agentId",
          "provider",
          "model",
          "promptVersion",
          "usage",
          "durationMs"
        ],
        "title": "LLMCallOut",
        "type": "object"
      },
...
      "PrivacyReportOut": {
        "properties": {
          "detections": {
            "items": {
              "$ref": "#/components/schemas/PhiDetectionOut"
            },
            "title": "Detections",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "minimumNecessary": {
            "default": "",
            "title": "Minimumnecessary",
            "type": "string"
          },
          "safeguards": {
            "items": {
              "$ref": "#/components/schemas/SafeguardCheckOut"
            },
            "title": "Safeguards",
            "type": "array"
          }
        },
        "title": "PrivacyReportOut",
        "type": "object"
      },
      "ProfileSnapshotOut": {
        "description": "Lo que se congeló al arrancar la ejecución. No cambia nunca.",
        "properties": {
          "maxSteps": {
            "default": 4,
            "title": "Maxsteps",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "promptVersion": {
            "title": "Promptversion",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "temperature": {
            "default": 0.0,
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "provider",
          "model",
          "promptVersion"
        ],
        "title": "ProfileSnapshotOut",
        "type": "object"
      },
...
      "RunDetailOut": {
        "properties": {
          "deliverables": {
            "$ref": "#/components/schemas/DeliverablesOut"
          },
          "events": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TraceEventOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events"
          },
          "run": {
            "$ref": "#/components/schemas/RunOut"
          }
        },
        "required": [
          "run",
          "deliverables"
        ],
        "title": "RunDetailOut",
        "type": "object"
      },
      "RunOut": {
        "properties": {
          "cancelledAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelledat"
          },
          "enabledAgents": {
            "items": {
              "type": "string"
            },
            "title": "Enabledagents",
            "type": "array"
          },
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "executions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentExecutionOut"
            },
            "title": "Executions",
            "type": "object"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "profiles": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ProfileSnapshotOut"
            },
            "title": "Profiles",
            "type": "object"
          },
          "requirementId": {
            "title": "Requirementid",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "title": "Startedat",
            "type": "string"
          },
          "startedBy": {
            "title": "Startedby",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageOut"
          }
        },
        "required": [
          "id",
          "requirementId",
          "startedBy",
          "startedAt",
          "status",
          "usage"
        ],
        "title": "RunOut",
        "type": "object"
      },
...
      "SafeguardCheckOut": {
        "properties": {
          "evidence": {
            "default": "",
            "title": "Evidence",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "title": "SafeguardCheckOut",
        "type": "object"
      },
...
      "SqlReportOut": {
        "properties": {
          "engine": {
            "default": "",
            "title": "Engine",
            "type": "string"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "scripts": {
            "items": {
              "$ref": "#/components/schemas/SqlScriptOut"
            },
            "title": "Scripts",
            "type": "array"
          }
        },
        "title": "SqlReportOut",
        "type": "object"
      },
...
      "TestsReportOut": {
        "properties": {
          "command": {
            "default": "",
            "title": "Command",
            "type": "string"
          },
          "coverage": {
            "default": 0.0,
            "title": "Coverage",
            "type": "number"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "framework": {
            "title": "Framework",
            "type": "string"
          },
          "tests": {
            "items": {
              "$ref": "#/components/schemas/GeneratedTestOut"
            },
            "title": "Tests",
            "type": "array"
          }
        },
        "required": [
          "framework"
        ],
        "title": "TestsReportOut",
        "type": "object"
      },
      "TraceEventOut": {
        "properties": {
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UsageOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "seq",
          "runId",
          "at",
          "type"
        ],
        "title": "TraceEventOut",
        "type": "object"
      },
...
      "UiuxReportOut": {
        "properties": {
          "baseUrl": {
            "default": "",
            "title": "Baseurl",
            "type": "string"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/UiScenarioOut"
            },
            "title": "Scenarios",
            "type": "array"
          }
        },
        "title": "UiuxReportOut",
        "type": "object"
      },
...
      "UsageOut": {
        "properties": {
          "costUsd": {
            "default": 0.0,
            "title": "Costusd",
            "type": "number"
          },
          "tokensIn": {
            "default": 0,
            "title": "Tokensin",
            "type": "integer"
          },
          "tokensOut": {
            "default": 0,
            "title": "Tokensout",
            "type": "integer"
          }
        },
        "title": "UsageOut",
        "type": "object"
      },
...
      "VerdictReportOut": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "guardrails": {
            "items": {
              "type": "string"
            },
            "title": "Guardrails",
            "type": "array"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "verdict",
          "confidence",
          "rationale"
        ],
        "title": "VerdictReportOut",
        "type": "object"
      },
      "VtrReportOut": {
        "properties": {
          "outputName": {
            "default": "",
            "title": "Outputname",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/VtrSectionOut"
            },
            "title": "Sections",
            "type": "array"
          },
          "templateName": {
            "default": "",
            "title": "Templatename",
            "type": "string"
          }
        },
        "title": "VtrReportOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend del orquestador multiagente. El dominio gira alrededor de un Requerimiento: adjuntos, plan de agentes, ejecuciones y entregables.",
    "title": "Orquestador de revisión de requerimientos",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "detail_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Incluir la traza completa",
            "in": "query",
            "name": "events",
            "required": false,
            "schema": {
              "default": false,
              "description": "Incluir la traza completa",
              "title": "Events",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Detail",
        "tags": [
          "ejecuciones"
        ]
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "description": "Detiene la ejecución: corta las llamadas en curso y cierra los agentes pendientes.",
        "operationId": "cancel_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Cancel",
        "tags": [
          "ejecuciones"
        ]
      }
    },
...
    "/api/runs/{run_id}/events": {
      "get": {
        "description": "Traza persistida. El canal en vivo con replay llega en ORQ-17.",
        "operationId": "trace_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "afterSeq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Afterseq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TraceEventOut"
                  },
                  "title": "Response Trace Api Runs  Run Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Trace",
        "tags": [
          "ejecuciones"
        ]
      }
    },
...
  }
}